fix warnings & fix comportement quand on supprime un leader et il y a un autre leader dispo

+++ TP-2/Buckland_Chapter3-Steering Behaviors/AgentLeader.cpp
@@ -84,5 +84,6 @@ void AgentLeader::getContextOfNewPursuer(Vehicle** leaderOfNewPursuer, Vector2D*
 //
 //------------------------------------------------------------------------
 AgentLeader::~AgentLeader() {
-	for (int i = 0; i < m_vAgentsPoursuiveurs.size(); i++) m_vAgentsPoursuiveurs[i]->stopFollowing();
+	for (size_t i = 0; i < m_vAgentsPoursuiveurs.size(); i++) m_vAgentsPoursuiveurs[i]->stopFollowing();
+	
 }
